# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "运维管理平台 API 文档",
        "title": "OpsHub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/monitor/alerts/channels": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertChannel"
                ],
                "summary": "获取告警通道列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AlertChannel"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertChannel"
                ],
                "summary": "创建告警通道",
                "parameters": [
                    {
                        "description": "告警通道信息",
                        "name": "channel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AlertChannel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AlertChannel"
                        }
                    }
                }
            }
        },
        "/monitor/alerts/channels/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertChannel"
                ],
                "summary": "获取告警通道详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "通道ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AlertChannel"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertChannel"
                ],
                "summary": "更新告警通道",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "通道ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "告警通道信息",
                        "name": "channel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AlertChannel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AlertChannel"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertChannel"
                ],
                "summary": "删除告警通道",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "通道ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/monitor/alerts/logs": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertLog"
                ],
                "summary": "获取告警日志列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "域名监控ID",
                        "name": "domainMonitorId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "告警类型",
                        "name": "alertType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AlertLog"
                            }
                        }
                    }
                }
            }
        },
        "/monitor/alerts/receiver-channels/:receiverId": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertReceiverChannel"
                ],
                "summary": "获取接收人的通道关联列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "接收人ID",
                        "name": "receiverId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AlertReceiverChannel"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertReceiverChannel"
                ],
                "summary": "添加接收人通道关联",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "接收人ID",
                        "name": "receiverId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "关联信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AlertReceiverChannel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AlertReceiverChannel"
                        }
                    }
                }
            }
        },
        "/monitor/alerts/receiver-channels/:receiverId/:channelId": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertReceiverChannel"
                ],
                "summary": "更新接收人通道关联配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "接收人ID",
                        "name": "receiverId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "通道ID",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "配置",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AlertReceiverChannel"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertReceiverChannel"
                ],
                "summary": "删除接收人通道关联",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "接收人ID",
                        "name": "receiverId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "通道ID",
                        "name": "channelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/monitor/alerts/receivers": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertReceiver"
                ],
                "summary": "获取告警接收人列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AlertReceiver"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertReceiver"
                ],
                "summary": "创建告警接收人",
                "parameters": [
                    {
                        "description": "告警接收人信息",
                        "name": "receiver",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AlertReceiver"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AlertReceiver"
                        }
                    }
                }
            }
        },
        "/monitor/alerts/receivers/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertReceiver"
                ],
                "summary": "获取告警接收人详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "接收人ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AlertReceiver"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertReceiver"
                ],
                "summary": "更新告警接收人",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "接收人ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "告警接收人信息",
                        "name": "receiver",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AlertReceiver"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AlertReceiver"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AlertReceiver"
                ],
                "summary": "删除告警接收人",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "接收人ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AlertChannel": {
            "type": "object",
            "properties": {
                "channelType": {
                    "description": "通道类型: email, webhook, wechat, dingtalk, feishu",
                    "type": "string"
                },
                "config": {
                    "description": "通道配置（JSON格式存储）",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "enabled": {
                    "description": "是否启用",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "通道名称",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.AlertReceiverChannel": {
            "type": "object",
            "properties": {
                "channelId": {
                    "description": "通道ID",
                    "type": "integer"
                },
                "config": {
                    "description": "在关联时使用的附加配置（JSON格式）\n用于存储通道特定的接收人配置，如飞书的user_id或phone等",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "receiverId": {
                    "description": "接收人ID",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}